Added one more test for the __getEmail__ method
which is really supposed to be used in only one place.
Renamed `__getEmail` to `__getEmail__`, otherwise, I can't access it from the test.
I am not quite sure why - it worked from `cfc_webapp.py`, but don't have time
to debug now, and it is not incorrect to have the method like this.