Remove all the obsolete "clients" code
Since we will use an ionic deploy-like solution for custom clients.
And even for ionic deploy, over the long-term, we don't necessarily want or
need people to check in UI code into our repository - instead, they should be
able to independently develop their UI, check it in to their own repository,
and to deploy to their own channel.