Make it easier to set up a well-defined server environment
Switch to `conda env` instead of pip for setting up the environment.
This makes it easier to include packages that are installed through conda (e.g.
`sklearn`, `numpy`, `pandas`, etc)

Add a simple script that can be run to set up the environment. Although it is a
two-line script, it makes it easier for us to run automated tests.

source the new environment before running the tests. Note that this is not
strictly necessary for manual execution, but again, helps with automated
testing.