Modify setup scripts to work with prefixes + teardown
Based on feedback from @shaneknapp, modify the test setup scripts to work with
prefixes. The test setup now installs to a temporary prefix location and tears
it down at the end. It also runs the other deployment scripts (e.g.
`bin/deploy/*`) so we can now have a single script to do all the setup.

The regular setup mirrors the test setup, but with a named environment and
without the additional deploy scripts.

Both setup scripts now source the environment that they just set up - the
source command was moved out of `runAllTests.sh` since the enviroments that it
runs in could have been set up in different ways.

Automated tests can now be run using:

```
$ source setup/setup_tests.sh
$ ./runAllTests.sh
$ source setup/teardown_tests.sh
```