Handle the case where the first point of the next section is returned in the query
This fixes
https://github.com/e-mission/e-mission-server/issues/530#issuecomment-352206464

Basically, two sections are back to back, then the last point of the first
section will overlap with the first point of the second section. So a query
based on the start and end time for the first section will return the the first
point of the second section as well, which causes a mismatch between the
re-retrieved and stored speeds and distances.

We detect and drop the last point in this case.