Switch to reading the list of all users from the uuid_db
Rationale:
- queries on timeseries are very slow
- we have entries in the timeseries, from stats conversion, that still have
  uuids from before the switch. We're keeping them around so that we can do
  analysis on the client timeseries alone, separated from

This implies that using the entries from the timeseries is not only slow but
incorrect. Let's just switching to reading entries from the uuid db. We've had
the client with the registration (to create uuid) forever - anything older is
probably incompatible anyway.

Let's also switch to the profile_db for platform specific queries. In this
case, the profile update was released months ago, so again, we expect that all
current clients should have updated by now.