Add special handling for the AIR mode
AIR is not a mode that is sensed from the phone, but it is inferred through
some heuristics in cleanAndResample instead of through the decision tree.
Ideally those heurstics should be replaced by the inference through the
decision tree, or through a separate heuristic step. But we are out of time for
a bigger refactor here.  so we say that if the sensed mode == AIR, we are going
to use it directly and ignore the inferred mode

Also add a new test case for this.

Also check for the inferred section, not just the inference result.
Had to add a new function to retrieve that, which involved refactoring some
existing code

+ Fixed AIR -> AIR_OR_HSR for the inferred modes as well